backend/pyproject.toml
[project]
name = "url-talk"
version = "0.1.0"
description = ""
authors = [
    {name = "Elton Costa",email = "89019551+ryzxxn@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "crawl4ai (>=0.6.3,<0.7.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "chromadb (>=1.0.12,<2.0.0)",
    "fastembed (>=0.7.0,<0.8.0)",
    "instructor[groq] (>=1.8.3,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "fitz (>=0.0.1.dev2,<0.0.2)",
    "docx (>=0.2.4,<0.3.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "playwright (>=1.57.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
